chore: move runGGShieldCommand to its own file
Issue: #
